Add #prev and #upcoming scopes to Event
First created as methods, then refactored to scopes.

Refactored _header.html.erb to have links to previous and upcoming
events. The links send the event type to the EventsController in
params[:events].

Refactored Events#index to trigger either the previous or upcoming
events scope, depending on what was sent in the params. If the user
navigates to '/events', upcoming events are used by default. In this
way, the index shows either previous or upcoming events, but not both.

Properly updated the application.js and gemfile to finish pulling in
bootstrap 4.

Updated Users#new view to the current theme.